pass index to replace function in emojify.replace
I am working on a React component that wraps emojify.js.

Having access to the index of the match would save me a bit of work,
since I need to break up the string into React elements to replace the
emojis.

A more complete solution would return an array of match objects, similar
to linkify-it.  But this is a simple change that gets me 90% there.

Thanks!